format :ゲッターとセッターの位置を修正
